FEAT/REF - Remove/Refactor actions/reducers to accept server playloads
