+++ app/reducers/game.js
@@ -1,14 +1,9 @@
-import { SOCKET_DEALER_RECEIVE_PROMPT } from '../action_types/actionTypes.js';
+import { UPDATE_GAME_STATE } from '../action_types/actionTypes.js';
 
 export default (state = {}, action = {}) => {
   switch (action.type) {
-    case SOCKET_DEALER_RECEIVE_PROMPT:
-      return Object.assign({}, state, {
-        dealerid: action.dealerid,
-        prompt: action.prompt,
-        category: action.category,
-        initialClue: action.initialClue,
-      });
+    case UPDATE_GAME_STATE:
+      return Object.assign({}, state, action.payload);
     default:
       return state;
   }
